pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[project]
authors = [
  { email = "30631553+liblaf@users.noreply.github.com", name = "liblaf" },
]
classifiers = ["Private :: Do Not Upload"]
dependencies = [
  "captum>=0.7.0,<0.8",
  "jaxtyping>=0.2.36,<0.3",
  "liblaf-toolkit>=0.0.11,<0.0.12",
  "matplotlib>=3.9.2,<4",
  "medmnist>=3.0.2,<4",
  "numpy>=2.1.2,<3",
  "opencv-python>=4.10.0.84,<5",
  "pandas>=2.2.3,<3",
  "polars>=1.15.0,<2",
  "pooch>=1.8.2,<2",
  "rich>=13.9.4,<14",
  "scikit-learn>=1.5.2,<2",
  "seaborn>=0.13.2,<0.14",
  "torch>=2.5.1,<3",
  "torchvision>=0.20.1,<0.21",
  "tqdm>=4.66.6,<5",
  "wandb>=0.18.5,<0.19",
]
description = "Add a short description here"
name = "deep-learning"
requires-python = ">=3.11,<3.13"
version = "0.0.0"

[project.optional-dependencies]
dev = ["icecream"]

[tool.pixi.environments]
default = ["dev"]

[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["linux-64"]
